changed checkPermissionAndDeleteOwnershipForObjectRemoval in favor of a type safe version that does not handle with strings

+++ java/com.sap.sse.security.ui/src/main/java/com/sap/sse/security/ui/server/UserManagementServiceImpl.java
@@ -126,14 +126,15 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
 
     @Override
     public void deleteRoleDefinition(String roleIdAsString) {
-        getSecurityService().checkPermissionAndDeleteOwnershipForObjectRemoval(SecuredSecurityTypes.ROLE_DEFINITION,
-                roleIdAsString, new Action() {
-                    @Override
-                    public void run() throws Exception {
-                        getSecurityService().deleteRoleDefinition(
-                                getSecurityService().getRoleDefinition(UUID.fromString(roleIdAsString)));
-                    }
-                });
+        RoleDefinition role = getSecurityService().getRoleDefinition(UUID.fromString(roleIdAsString));
+        if (role != null) {
+            getSecurityService().checkPermissionAndDeleteOwnershipForObjectRemoval(role, new Action() {
+                @Override
+                public void run() throws Exception {
+                    getSecurityService().deleteRoleDefinition(role);
+                }
+            });
+        }
     }
 
     @Override
@@ -325,22 +326,21 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
 
     @Override
     public SuccessInfo deleteUserGroup(String userGroupIdAsString) throws UnauthorizedException {
-        try {
-            return getSecurityService().checkPermissionAndDeleteOwnershipForObjectRemoval(SecuredSecurityTypes.USER_GROUP,
-                    userGroupIdAsString, () -> {
-                        try {
-                            final UUID userGroupId = UUID.fromString(userGroupIdAsString);
-                            final UserGroup userGroup = getSecurityService().getUserGroup(userGroupId);
-                            getSecurityService().deleteUserGroup(userGroup);
-                            return new SuccessInfo(true, "Deleted user group: " + userGroup.getName() + ".",
-                                    /* redirectURL */ null, null);
-                        } catch (UserGroupManagementException e) {
-                            return new SuccessInfo(false, "Could not delete user group.", /* redirectURL */ null, null);
-                        }
-                    });
-        } catch (AuthorizationException e) {
-            return new SuccessInfo(false, "You are not permitted to delete user group " + userGroupIdAsString, /* redirectURL */ null,
-                    null);
+        final UUID userGroupId = UUID.fromString(userGroupIdAsString);
+        final UserGroup userGroup = getSecurityService().getUserGroup(userGroupId);
+        if (userGroup != null) {
+            return getSecurityService().checkPermissionAndDeleteOwnershipForObjectRemoval(userGroup, () -> {
+                try {
+
+                    getSecurityService().deleteUserGroup(userGroup);
+                    return new SuccessInfo(true, "Deleted user group: " + userGroup.getName() + ".",
+                            /* redirectURL */ null, null);
+                } catch (UserGroupManagementException e) {
+                    return new SuccessInfo(false, "Could not delete user group.", /* redirectURL */ null, null);
+                }
+            });
+        } else {
+            return new SuccessInfo(false, "Could not delete user group.", /* redirectURL */ null, null);
         }
     }
 
@@ -636,26 +636,30 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
 
     @Override
     public SuccessInfo deleteUser(String username) throws UnauthorizedException {
-        // In addition to checking the default delete permission we currently explicitly require admin role for the owning user/group
-        if (!getSecurityService().hasCurrentUserRoleForOwnedObject(SecuredSecurityTypes.USER, username,
-                AdminRole.getInstance())) {
-            return new SuccessInfo(false, "You are not permitted to delete user " + username, /* redirectURL */ null,
-                    null);
-        }
-        try {
-            return getSecurityService().checkPermissionAndDeleteOwnershipForObjectRemoval(SecuredSecurityTypes.USER,
-                    username, () -> {
-                        try {
-                            getSecurityService().deleteUser(username);
-                            return new SuccessInfo(true, "Deleted user: " + username + ".", /* redirectURL */ null,
-                                    null);
-                        } catch (UserManagementException e) {
-                            return new SuccessInfo(false, "Could not delete user.", /* redirectURL */ null, null);
-                        }
-                    });
-        } catch (AuthorizationException e) {
-            return new SuccessInfo(false, "You are not permitted to delete user " + username, /* redirectURL */ null,
-                    null);
+        User user = getSecurityService().getUserByName(username);
+        if (user != null) {
+            // In addition to checking the default delete permission we currently explicitly require admin role for the
+            // owning user/group
+            if (!getSecurityService().hasCurrentUserRoleForOwnedObject(SecuredSecurityTypes.USER, username,
+                    AdminRole.getInstance())) {
+                return new SuccessInfo(false, "You are not permitted to delete user " + username,
+                        /* redirectURL */ null, null);
+            }
+            try {
+                return getSecurityService().checkPermissionAndDeleteOwnershipForObjectRemoval(user, () -> {
+                    try {
+                        getSecurityService().deleteUser(username);
+                        return new SuccessInfo(true, "Deleted user: " + username + ".", /* redirectURL */ null, null);
+                    } catch (UserManagementException e) {
+                        return new SuccessInfo(false, "Could not delete user.", /* redirectURL */ null, null);
+                    }
+                });
+            } catch (AuthorizationException e) {
+                return new SuccessInfo(false, "You are not permitted to delete user " + username,
+                        /* redirectURL */ null, null);
+            }
+        } else {
+            return new SuccessInfo(false, "Could not delete user.", /* redirectURL */ null, null);
         }
     }
